+++ src/pages/main-page.tsx
@@ -4,7 +4,7 @@ import RestCardWrapper from '../UI/rest-card-wrapper/rest-card-wrapper';
 import ProgressBar from '../UI/progress-bar/progress-bar';
 import SearchNotification from '../UI/search-notification/search-notification';
 import { useStoreSearchQuery, useStoreSearchResult } from '../assets/functions/hooks/redux.hooks';
-import { useGetAuthorsByNameQuery } from '../services/openLibraryApi';
+import { useGetAuthorsByNameQuery } from '../services/libraryApi';
 import { getRTKFetchErrorMessage } from '../assets/functions/handlers/error-handler-functions';
 import { RestAuthorData } from 'interfaces';
